add smart components
